Struct nekoton_abi::transaction_parser::TransactionParserBuilder

pub struct TransactionParserBuilder { /* private fields */ }

Implementations

impl TransactionParserBuilder

pub fn function_input( self, function: Function, allow_partial_match: bool ) -> Self

Matches all messages with in function_id.

Params:

allow_partial_match - if true, won’t return error if there are unparsed bytes left in the message. Set false by default.

pub fn function_output( self, function: Function, allow_partial_match: bool ) -> Self

Matches all messages with out function_id

Params:

allow_partial_match - if true, won’t return error if there are unparsed bytes left in the message. Set false by default.

pub fn function_bounce(self, function: FunctionWithBounceHandler) -> Self

Matches in messages with function_id and applies bounce handler

pub fn event(self, event: Event) -> Self

Matches out messages with event_id

pub fn function_in_list<I>( self, functions: I, allow_partial_match: bool ) -> Selfwhere I: IntoIterator<Item = Function>,

pub fn functions_out_list<I>( self, functions: I, allow_partial_match: bool ) -> Selfwhere I: IntoIterator<Item = Function>,

pub fn events_list<I>(self, events: I) -> Selfwhere I: IntoIterator<Item = Event>,

pub fn build(self) -> Result<TransactionParser>

Returns Err if there are duplicate function_ids

pub fn build_with_external_in(self) -> Result<TransactionParser>

Returns Err if there are duplicate function_ids or different contracts are used