Crate nekoton_abi

Re-exports

pub use self::transaction_parser::TransactionParser;

pub use self::tvm::BriefBlockchainConfig;

pub use self::tvm::StackItem;

pub use self::tvm::VmGetterOutput;

pub use num_bigint;

pub use num_traits;

Modules

address_only_hash

array_address_only_hash

array_int256_number

array_uint160_bytes

bytes_as_string

int256_number

transaction_parser

tvm

uint128_number

uint160_bytes

uint256_number

Structs

BigUint128

BigUint256

ContractOutputUnpacker

EventBuilder

ExecutionContext

ExecutionOutput

Executor

ExpireHeader

FunctionBuilder

MaybeRef

TokenValue::Optional which always store its value in the cell

MessageBuilder

PubkeyHeader

TimeHeader

TransactionId

TupleBuilder

Enums

ExtractionError

GenTimings

LastTransactionId

MethodName

TokensJsonError

UnpackerError

Constants

GEN_TIMINGS_ALLOWABLE_INTERVAL

Additional estimated lag for the pending message to be expired

Traits

AsGetterMethodId

BuildMapKeyTokenValue

BuildTokenValue

FunctionExt

FunctionOutputMarker

IntoUnpacker

KnownParamType

KnownParamTypeArray

KnownParamTypePlain

PackAbi

PackAbiPlain

StandaloneToken

TokenValueExt

UnpackAbi

UnpackAbiPlain

UnpackFirst

UnpackHeader

Functions

answer_id

code_to_tvc

create_boc_or_comment_payload

Tries to parse text as boc, encodes as comment otherwise

create_boc_payload

Creates slice data from base64 encoded boc

create_comment_payload

Creates slice data with string, encoded as comment

decode_event

decode_input

decode_output

default_blockchain_config

extract_public_key

get_code_salt

get_state_init_hash

guess_method_by_input

insert_state_init_data

make_abi_token_value

make_abi_tokens

pack_into_cell

parse_abi_token

parse_abi_token_value

parse_abi_tokens

parse_comment_payload

parse_transaction_messages

process_out_messages

process_raw_outputs

read_function_id

read_input_function_id

set_code_salt

unpack_from_cell

unpack_headers

Type Aliases

UnpackerResult

Derive Macros

KnownParamType

KnownParamTypePlain

PackAbi

PackAbiPlain

UnpackAbi

UnpackAbiPlain